showing all info on index

+++ src/services/bio/index.ts
@@ -19,11 +19,20 @@ export const BioService = {
 
         const userBio = await Bio.create({ user, bio })
 
+        hasUser.bio = userBio
+        await hasUser.save()
+
         return formatResponse(201, userBio)
     },
     async delete(user: string) {
+        const hasUser = await User.findById(user)
+        if (!hasUser) return formatResponse(404, 'User does not exist')
+
         const deletedBio = await Bio.findOneAndRemove({ user: user })
         if (!deletedBio) return formatResponse(404, 'Bio did not exist')
+
+        hasUser.bio = undefined
+        await hasUser.save()
 
         return formatResponse(200, { deleted: 'ok', deletedBio })
     }
